requirements.txt
Nuitka==2.5.1
numpy==2.2.3
ordered-set==4.1.0
pandas==2.2.3
patchelf==0.17.2.2
peewee==3.17.9
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
shiboken6==6.8.2.1
six==1.17.0
tzdata==2025.1
zstandard==0.23.0
